Move base dependency check to a task
